Moving config files to config dir
